This patch extends ginger capabilities to do listing/format of DASD devices and listing/create/delete of DASD Partitions on host
Adding DASD devices list/format, DASDPartitions list/create/delete support to Ginger

This patch adds a Collection 'dasdevs' and a Resource
'dasddev' to retrieve and manipulate the DASD devices of the host.
This patch also adds a Collection 'dasdpartitions' and a Resource
'dasdpartition' to retrieve and manipulate the DASD Partitions.

Test cases for DASD support
Includes tests of listing DASD devices and parser of 'lsdasd -l' command output.
Includes tests of listing DASD Partitions and create a DASD partition.